Refactor authentication into SilverfinAuthorizer class
