+++ lib/api/sfApi.js
@@ -1,57 +1,20 @@
-const axios = require("axios");
-const open = require("open");
-const prompt = require("prompt-sync")({ sigint: true });
 const apiUtils = require("../utils/apiUtils");
 const { consola } = require("consola");
 const { firmCredentials } = require("../api/firmCredentials");
 const { AxiosFactory } = require("./axiosFactory");
+const { SilverfinAuthorizer } = require("./silverfinAuthorizer");
 
 apiUtils.checkRequiredEnvVariables();
 
-async function authorizeApp(firmId = undefined) {
-  try {
-    consola.info(
-      `NOTE: if you need to exit this process you can press "Ctrl/Cmmd + C"`
-    );
-    const redirectUri = encodeURIComponent("urn:ietf:wg:oauth:2.0:oob");
-    const scope = encodeURIComponent(
-      "administration:read administration:write financials:read financials:write workflows:read"
-    );
-
-    let firmIdPrompt;
-    if (firmId) {
-      firmIdPrompt = prompt(
-        `Enter the firm ID (leave blank to use ${firmId}): `,
-        { value: firmId }
-      );
-    } else {
-      firmIdPrompt = prompt("Enter the firm ID: ");
-    }
-    const url = `${apiUtils.BASE_URL}/f/${firmIdPrompt}/oauth/authorize?client_id=${process.env.SF_API_CLIENT_ID}&redirect_uri=${redirectUri}&response_type=code&scope=${scope}`;
-
-    await open(url);
-    consola.info(`You need to authorize your APP in the browser`);
-    consola.log("Insert your credentials...");
-    const authCodePrompt = prompt("Enter your API authorization code: ", {
-      echo: "*",
-    });
-    // Get tokens
-    const tokens = await apiUtils.getAccessToken(firmIdPrompt, authCodePrompt);
-    if (tokens) {
-      consola.success("Authentication successful");
-    }
-    // Get firm name
-    await apiUtils.getFirmName(firmIdPrompt);
-  } catch (error) {
-    consola.error(error);
-    process.exit(1);
-  }
+async function authorizeFirm(firmId) {
+  SilverfinAuthorizer.authorizeFirm(firmId);
 }
 
 async function refreshFirmTokens(firmId) {
   try {
     const instance = AxiosFactory.createInstance("firm", firmId);
     const firmTokens = firmCredentials.getTokenPair(firmId);
+    const BASE_URL = firmCredentials.getHost();
     let data = {
       client_id: process.env.SF_API_CLIENT_ID,
       client_secret: process.env.SF_API_SECRET,
@@ -61,7 +24,7 @@ async function refreshFirmTokens(firmId) {
       access_token: firmTokens.accessToken,
     };
     const response = await instance.post(
-      `${apiUtils.BASE_URL}/f/${firmId}/oauth/token`,
+      `${BASE_URL}/f/${firmId}/oauth/token`,
       data
     );
 
@@ -92,9 +55,10 @@ async function refreshPartnerToken(partnerId) {
       process.exit(1);
     }
 
+    const BASE_URL = firmCredentials.getHost();
     const instance = AxiosFactory.createInstance("partner", partnerId);
     const response = await instance.post(
-      `${apiUtils.BASE_URL}/api/partner/v1/refresh_api_key?api_key=${partnerCredentials.token}`
+      `${BASE_URL}/api/partner/v1/refresh_api_key?api_key=${partnerCredentials.token}`
     );
 
     firmCredentials.storePartnerApiKey(
@@ -838,7 +802,7 @@ async function getFirmDetails(firmId) {
 }
 
 module.exports = {
-  authorizeApp,
+  authorizeFirm,
   refreshFirmTokens,
   refreshPartnerToken,
   createReconciliationText,
