+++ lib/utils/apiUtils.js
@@ -1,8 +1,5 @@
 const { firmCredentials } = require("../api/firmCredentials");
-const axios = require("axios");
 const { consola } = require("consola");
-
-const BASE_URL = firmCredentials.getHost();
 
 function checkAuthorizePartners(partner_id) {
   const partnerCredentials = firmCredentials.getPartnerCredentials(partner_id);
@@ -20,32 +17,6 @@ function checkRequiredEnvVariables() {
     consola.log(`Call export ${missingVariables[0]}=...`);
     consola.log(
       `If you don't have credentials yet, you need to register your app with Silverfin to get them`
-    );
-    process.exit(1);
-  }
-}
-
-// Get Tokens for the first time with an authorization code
-async function getAccessToken(firmId, authCode) {
-  try {
-    const redirectUri = "urn%3Aietf%3Awg%3Aoauth%3A2.0%3Aoob";
-    const grantType = "authorization_code";
-    let requestDetails = {
-      method: "POST",
-      url: `${BASE_URL}/f/${firmId}/oauth/token?client_id=${process.env.SF_API_CLIENT_ID}&client_secret=${process.env.SF_API_SECRET}&redirect_uri=${redirectUri}&grant_type=${grantType}&code=${authCode}`,
-    };
-    const response = await axios(requestDetails);
-    firmCredentials.storeNewTokenPair(firmId, response.data);
-    await getFirmName(firmId);
-    return true;
-  } catch (error) {
-    consola.error(
-      `Response Status: ${error.response.status} (${error.response.statusText})`
-    );
-    consola.error(
-      `Error description: ${JSON.stringify(
-        error.response.data.error_description
-      )}`
     );
     process.exit(1);
   }
@@ -108,25 +79,9 @@ async function responseErrorHandler(error) {
   throw error;
 }
 
-/**
- * Retrieve firm details and store the firm name in the credentials file
- * @param {Number} firmId
- */
-async function getFirmName(firmId) {
-  try {
-    const response = await axios.get(`/user/firm`);
-    if (response && response.data) {
-      firmCredentials.storeFirmName(firmId, response.data.name);
-    }
-  } catch (error) {}
-}
-
 module.exports = {
-  BASE_URL,
+  checkAuthorizePartners,
   checkRequiredEnvVariables,
-  getAccessToken,
   responseSuccessHandler,
   responseErrorHandler,
-  getFirmName,
-  checkAuthorizePartners,
 };
